Moved all digest indices functions to a single file.
