+++ molsysmt/_private/digestion/__init__.py
@@ -1,5 +1,5 @@
 from .argument import digest_argument
-from .atom_indices import digest_atom_indices
+# from .atom_indices import digest_atom_indices
 from .box import digest_box
 from .box import digest_box_lengths, digest_box_lengths_value
 from .box import digest_box_angles, digest_box_angles_value
@@ -8,14 +8,15 @@ from .comparison import digest_comparison
 from .element import digest_element
 from .engine import digest_engine
 from .form import digest_form, digest_to_form
-from .group_indices import digest_group_indices
-from .indices import digest_indices
+# from .group_indices import digest_group_indices
+from .indices import digest_indices, digest_structure_indices, digest_multiple_structure_indices
+from .indices import digest_atom_indices, digest_group_indices
 from .item import digest_item
 from .molecular_system import digest_single_molecular_system, digest_multiple_molecular_systems
 from .output import digest_output
 from .selection import digest_selection, digest_multiple_selections
 from .step import digest_step
-from .structure_indices import digest_structure_indices, digest_multiple_structure_indices
+# from .structure_indices import digest_structure_indices, digest_multiple_structure_indices
 from .syntaxis import digest_syntaxis, digest_to_syntaxis
 from .element import digest_element
 from .time import digest_time
