--- molsysmt/_private/digestion/atom_indices.py
@@ -1,7 +1,0 @@
-from .indices import digest_indices
-
-
-def digest_atom_indices(atom_indices):
-    """ Checks if atom indices are of the expected type and value. """
-    return digest_indices(atom_indices)
-
