--- molsysmt/_private/digestion/group_indices.py
@@ -1,6 +1,0 @@
-import numpy as np
-from .indices import digest_indices
-
-def digest_group_indices(group_indices):
-    return digest_indices(group_indices)
-
